+++ application/src/test/java/org/opentripplanner/apis/gtfs/GraphQLIntegrationTest.java
@@ -108,7 +108,6 @@ import org.opentripplanner.transfer.regular.TransferServiceTestFactory;
 import org.opentripplanner.transit.model._data.TransitRepositoryForTest;
 import org.opentripplanner.transit.model.basic.Money;
 import org.opentripplanner.transit.model.basic.TransitMode;
-import org.opentripplanner.transit.model.calendar.DefaultTripCalendars;
 import org.opentripplanner.transit.model.framework.AbstractBuilder;
 import org.opentripplanner.transit.model.framework.Deduplicator;
 import org.opentripplanner.transit.model.network.BikeAccess;
@@ -300,13 +299,13 @@ class GraphQLIntegrationTest {
       cal_id,
       List.of(firstDate, secondDate, SERVICE_DATE)
     );
-    transitRepository.getServiceCodes().put(cal_id, SERVICE_CODE);
+    transitRepository.putServiceCode(cal_id, SERVICE_CODE);
     transitRepository.updateCalendarServiceData(calendarServiceData);
     transitRepository.index();
 
     DefaultTimetableRepository timetableSnapshot = new DefaultTimetableRepository(
       RaptorTransitDataTestFactory.empty(),
-      new DefaultTripCalendars()
+      transitRepository.getTripCalendar()
     );
     timetableSnapshot.update(
       RealTimeTripUpdate.of(
@@ -473,6 +472,14 @@ class GraphQLIntegrationTest {
       .withDescriptionText(I18NString.of("This station is currently closed"))
       .withEffect(AlertEffect.NO_SERVICE)
       .addEntity(stationEntitySelector)
+      // deliberately unsorted and with open bounds to test the sorting of the activity periods
+      .withCalendar(
+        AlertCalendar.of(
+          TimePeriod.of(ALERT_END_TIME, null),
+          TimePeriod.of(ALERT_START_TIME, ALERT_END_TIME),
+          TimePeriod.of(null, ALERT_START_TIME)
+        )
+      )
       .build();
 
     // TODO - Use itineraryBuilder() here not build() and complete building the itinerary using